pyproject.toml
[project]
name = "nutrition-tracker-bot"
version = "0.1.0"
description = "Telegram bot for personal nutrition and activity tracking"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.26.0",
    "aiohttp>=3.11.0",
    "aiosqlite>=0.21.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "SpeechRecognition>=3.14.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
